refactor(frontend): extract upload files to service handler
